Replace transition with css animation
This would solve a problem that can happen, especially if the images already have a transition effect applied. With animation the style is applied in a specific time frame without interfering with the css style. 

it's a small modification I know, but it can make everything more beautiful (or less glitchy). Thanks for your script it's very useful to me!